Update pannellum.css
This is an example where the image is set by default to 70% size, and the hover to 100% (using `CSS transitions`)
There is also an example of `CSS Keyframes animation` commented.

The right way to do this (not hardcoded) would be to *provide a [LESS](http://lesscss.org/) file, with the `initialWidth, hoverWidth and animationKeyFrameName` as variables*. 
Then we could provide diferent animation keyframes (which is used for the variable animationKeyFrameName) in the  in a separate css file. For example a fade animation, a width animatino, a change_color animation.