Prevent empty values in URL config from stopping panorama rendering
Should the user use URL configuration for their panorama and then
specify an option/key but leave the value empty, previously this
would stop the panorama from being displayed due to the generated
JSON config from src/standalone/standalone.js being invalid.

This commit omits creating an on-the-fly JSON string and instead
uses a JavaScript object (which the generated JSON string was
converted to anyway).

The background of wanting to enable URL options with possibly empty
values relates to using Pannellum with static site generators.
Depending on the static site generator, specifying default values
for undefined template variables can result in lots of template
notation whereas outputting a value for a variable that isn't
defined usuallly simply outputs an empty string. It would hence be
desirable to treat the empty-string value of an option/key in URL
config as if that option/key were not user-supplied at all and
hence the Pannellum default value for that option/key be used.