Fix `animateMove`, if any of axes is not changed.
It never stops calling `animateMove`, if `pitch`, `yaw`, or `hfov` has not been changed.
Since we may have callback function, `animateMove` has to be called anyway, even if we don't actually need to animate movement. That's why equality of start and end positions is checked here, but not in `setPitch`, `setYaw`, and `setHfov`.