Optimize multires tile display.
Previously, a single multires tile level was used for the entire viewport,
which resulted in too much resolution at the center and too little at the
edges. This was particularly noticable when zoomed out all the way.

The new approach selectively loads tiles based on their individual projected
resolution, which is more efficient and should resolve long-standing
edge-softness issues.