Use a list of missing multires tiles to avoid trying load them.
This avoids the 404 errors that would previously result.

A compact encoding is used, with `!` plus the face letter used to specify a
new face, `>` plus a base83-encoded number used to specify a new level, and a
list of base83-encoded numbers used to specify x and y tile coordinates, where
the base83 encoding uses the minimum number of characters to encode the
maximum tile number for the level in question. The base83 encoding is the same
as the one used for SHT hashes.